requirements.txt
streamlit
pandas
numpy
matplotlib
plotly.express
seaborn
pyspark
delta-spark
pyarrow
pydrive2
duckdb
functools